requirements.txt
Cerberus==1.3
Click==7.0
cycler==0.10.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
pymongo==3.7.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
six==1.12.0
Werkzeug==0.15.2
